EmoVoice/requirements.txt
absl-py==2.1.0
accelerate==0.34.2
addict==2.4.0
aiofiles==23.2.1
anyio==4.6.2.post1
asttokens==2.4.1
async-timeout==4.0.3
attrs==24.2.0
bitstring==4.2.3
blinker==1.8.2
Brotli==1.1.0
cachetools==4.2.4
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
conformer==0.3.2
cycler==0.12.1
Cython==3.0.11
datasets==3.3.2
debugpy==1.8.8
decorator==5.1.1
deepspeed==0.14.2
diffusers==0.27.2
dill==0.3.8
editdistance==0.8.1
einops==0.8.0
exceptiongroup==1.2.2
executing==2.1.0
fastapi==0.111.0
ffmpy==0.4.0
filelock==3.16.1
fire==0.6.0
frozenlist==1.4.1
fsspec==2024.2.0
gdown==5.1.0
gitdb==4.0.11
GitPython==3.1.43
huggingface-hub==0.25.2
hydra-core==1.3.2
HyperPyYAML==1.2.2
importlib_metadata==8.5.0
jsonschema==4.23.0
kaldiio==2.18.0
librosa==0.10.2.post1
lightning==2.2.4
lightning-utilities==0.11.8
loralib==0.1.2
Mako==1.3.5
Markdown==3.7
matplotlib==3.7.5
modelscope==1.18.1
more-itertools==10.5.0
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
multiprocess==0.70.16
networkx==3.1
ninja==1.11.1.1
numba==0.60.0
numpy==1.26.4
oauthlib==3.2.2
omegaconf==2.3.0
onnx==1.16.0
onnxruntime-gpu==1.16.0
openai-whisper
orjson==3.10.11
packaging==24.2
pandas==2.2.2
protobuf==4.25.5
pydantic==2.7.0
pydantic_core==2.18.1
pydash==8.0.3
pydub==0.25.1
Pygments==2.18.0
PyJWT==2.9.0
pynini==2.1.5
pynndescent==0.5.13
pynvml==11.5.3
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.17
pytorch-lightning==2.4.0
pytorch-wpe==0.0.1
pytz==2024.2
PyYAML==6.0.2
rdflib==7.1.3
referencing==0.35.1
regex==2024.9.11
requests==2.32.3
rotary-embedding-torch==0.8.6
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
segments==2.3.0
sentence-transformers==4.0.1
sentencepiece==0.2.0
soundfile==0.12.1
stack-data==0.6.3
starlette==0.37.2
tensorboard==2.14.0
tensorboardX==2.6.2.2
tokenizers==0.19.1
torch==2.4.1
torchaudio==2.4.1
torchmetrics==1.6.0
torchvision==0.19.1
tqdm==4.66.5
transformers==4.43.4
triton==3.0.0
typeguard==4.4.1
typer==0.13.0
typing_extensions==4.12.2
ujson==5.10.0
urllib3==2.2.3
uvicorn==0.30.0
uvloop==0.21.0
wandb==0.18.1
websockets==11.0.3
WeTextProcessing==1.0.3
wget==3.2
whisper_normalizer==0.0.10
wrapt==1.16.0
zipp==3.21.0
peft
funasr
